requirements.txt
flask
gunicorn
flask-cors
pywhispercpp
transformers
argostranslate
fpdf
python-docx
reportlab
